+++ src/shared/ui/calendar/utils/getDayName.ts
@@ -1,4 +1,4 @@
-import { WEEK_DAYS } from "../dayCell/constants"
+import { WEEK_DAYS } from "../dayCell/DayCell.constants"
 
 export const getDayName = (date: Date) => {
   const day = date.getDay()
